pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = ""
authors = ["Your Name <email@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = []

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.28.0"
casers = "^0.6.0"
fastapi = "^0.100"
orjson = "^3.8.3"
parol = "^1.0.1"
pydantic = "^2"
pydantic-settings = "^2.0.1"
pyjwt = {version = "^2.7.0", extras = ["crypto"]}
python-json-logger = "^2.0.7"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.18"}
typer = "^0.9.0"
uvicorn = {extras = ["standard"], version = "^0.22"}

[tool.poetry.group.dev.dependencies]
black = ">=23"
codecov = "^2"
httpx = "^0.24"
isort = "^5.12.0"
mypy = "^1.4"
poethepoet = "^0.20"
pytest = "^7"
pytest-asyncio = "*"
pytest-cov = "*"
pytest-deadfixtures = "*"
pytest-mock = "*"
ruff = "^0.0.277"

[tool.poetry.group.migration.dependencies]
alembic = "^1.9.1"
psycopg2-binary = "^2.9.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.env]
CODE = "tests app"

[tool.poe.tasks.all]
help = "Run format lint test"
sequence = ["format", "lint", "test"]
default_item_type = "ref"

[tool.poe.tasks.test]
help = "Test with coverage"
cmd = "pytest -vv --cov=./"

[tool.poe.tasks.test-fast]
help = "Test until error"
cmd = "pytest -vv --exitfirst"

[tool.poe.tasks.test-failed]
help = "Test failed"
cmd = "pytest -vv --last-failed"

[tool.poe.tasks.test-report]
help = "Report testing"
sequence = [
    {cmd = "pytest -vv --cov --cov-report html"},
    {cmd = "python -m webbrowser 'htmlcov/index.html'"},
]

[tool.poe.tasks.lint]
help = "Check code"
sequence = [
    {cmd = "ruff check ${CODE}"},
    {cmd = "black --check ${CODE}"},
    {cmd = "pytest --dead-fixtures --dup-fixtures"},
    {cmd = "mypy ${CODE}"},
]

[tool.poe.tasks.format]
help = "Formating code"
sequence = [
	{cmd = "ruff check --fix-only ${CODE}"},
	{cmd = "isort ${CODE}"},
	{cmd = "black ${CODE}"},
]

[tool.poe.tasks.migrate]
help = "Run the migration"
cmd = "alembic upgrade head"

[tool.poe.tasks.migration]
help = "Create new migration"
args   = [{ name = "message", positional = true, required = true }]
cmd = "alembic revision --autogenerate --message ${message}"

[tool.poe.tasks.cli]
help = "CLI"
cmd = "python -m app"

[tool.poe.tasks.run]
help = "Run"
cmd = "python -m app server up"

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.isort]
profile = "black"

[tool.coverage.run]
omit = ["tests/*", "**/__main__.py", "**/.venv/*", "**/site-packages/*"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 100

[tool.black]
skip-string-normalization = true

[tool.pytest.ini_options]
asyncio_mode = "strict"
addopts = "--strict-markers --showlocals --verbosity 2"
log_level = "DEBUG"

[tool.ruff]
line-length = 88
select = [
    "D",
    "B",
    "C",
    "E",
    "F",
    "N",
    "Q",
    "RUF001",
    "T",
    "W",
]
ignore = [
    "Q000",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D203",
    "D213",
    "B008",
]
exclude = []
